INTGW-435 Add extra filter option,Refactoring code
